+++ hybrid_core/hybrid_core/example/lib/view/os_view.dart
@@ -2,24 +2,32 @@ import 'package:flutter/material.dart';
 import 'package:hybrid_core/hybrid_core.dart';
 
 class OSView extends StatelessWidget {
-  const OSView({super.key});
+  final OS os;
+
+  OSView({super.key}) : os = OS();
 
   @override
   Widget build(BuildContext context) {
-    final os = OS();
+    final os = this.os;
     final String version;
     final bool atLeastVersion;
     if (os is Android) {
-      version = '${os.api}';
+      version = 'Android API${os.api}';
       atLeastVersion = os.atLeastAPI(33);
     } else if (os is iOS) {
-      version = '${os.version}';
+      version = 'iOS ${os.version}';
       final version17_0 = DarwinVersion.number(17.0);
       atLeastVersion = os.atLeastVersion(version17_0);
     } else if (os is macOS) {
-      version = '${os.version}';
+      version = 'macOS ${os.version}';
       final version14_0 = DarwinVersion.number(14.0);
       atLeastVersion = os.atLeastVersion(version14_0);
+    } else if (os is Windows) {
+      version = 'Windows ${os.version}';
+      atLeastVersion = os.isWindows10OrGreater;
+    } else if (os is Linux) {
+      version = 'Linux <UNKNOWN>';
+      atLeastVersion = false;
     } else {
       throw TypeError();
     }
@@ -27,7 +35,9 @@ class OSView extends StatelessWidget {
       appBar: AppBar(
         title: const Text('OS'),
       ),
-      body: Center(
+      body: Container(
+        margin: const EdgeInsets.all(20.0),
+        alignment: Alignment.center,
         child: Column(
           mainAxisSize: MainAxisSize.min,
           children: [
@@ -43,8 +53,9 @@ class OSView extends StatelessWidget {
               height: 20.0,
             ),
             Text(
-              'At least API 33+ or iOS 17.0+ or macOS 14.0+',
+              'Version at least\nAndroid API 33+\niOS 17.0+\nmacOS 14.0+\nWindows 10+',
               style: Theme.of(context).textTheme.bodyMedium,
+              textAlign: TextAlign.center,
             ),
             Text(
               '$atLeastVersion',