+++ hybrid_core/hybrid_core/lib/hybrid_core.dart
@@ -1,2 +1,10 @@
 export 'package:hybrid_core_platform_interface/hybrid_core_platform_interface.dart'
     hide OSPlatform;
+export 'package:hybrid_core_android/hybrid_core_android.dart'
+    hide HybridCoreAndroid, AndroidPlatform;
+export 'package:hybrid_core_darwin/hybrid_core_darwin.dart'
+    hide HybridCoreDarwin, DarwinPlatform, iOSPlatform, macOSPlatform;
+export 'package:hybrid_core_windows/hybrid_core_windows.dart'
+    hide HybridCoreWindows, WindowsPlatform;
+export 'package:hybrid_core_linux/hybrid_core_linux.dart'
+    hide HybridCoreLinux, LinuxPlatform;