+++ hybrid_core/hybrid_core_android/lib/src/os.dart
@@ -2,14 +2,7 @@ import 'package:hybrid_core_platform_interface/hybrid_core_platform_interface.da
 
 import 'jni.dart' as jni;
 
-class OSPlatformImpl extends OSPlatform {
-  @override
-  final OS os;
-
-  OSPlatformImpl() : os = AndroidImpl();
-}
-
-class AndroidImpl implements Android {
+class AndroidPlatform extends OSPlatform implements Android {
   @override
   int get api => jni.Build_VERSION.SDK_INT;
 
@@ -17,4 +10,10 @@ class AndroidImpl implements Android {
   bool atLeastAPI(int api) {
     return jni.Build_VERSION.SDK_INT >= api;
   }
+}
+
+abstract class Android implements OS {
+  int get api;
+
+  bool atLeastAPI(int api);
 }