+++ hybrid_core/hybrid_core_darwin/lib/hybrid_core_darwin.dart
@@ -1,9 +1,2 @@
-import 'package:hybrid_core_platform_interface/hybrid_core_platform_interface.dart';
-
-import 'src/os.dart';
-
-abstract class HybridCoreDarwin {
-  static void registerWith() {
-    OSPlatform.instance = OSPlatformImpl();
-  }
-}
+export 'src/hybrid_core_darwin.dart';
+export 'src/os.dart';