+++ hybrid_core/hybrid_core_android/CHANGELOG.md
@@ -9,3 +9,9 @@
 ## 0.1.0
 
 * Move `AndroidOSImpl` to `AndroidImpl`.
+
+## 0.2.0
+
+* Add `Android` class.
+* Move `OSPlatformImpl` to `AndroidPlatform`.
+* Remove `AndroidImpl` class.