+++ hybrid_core/hybrid_core_darwin/CHANGELOG.md
@@ -13,3 +13,15 @@
 ## 0.1.0
 
 * Move `DarwinOSImpl` to `DarwinImpl`.
+
+## 0.2.0
+
+* Add `DarwinPlatform` class.
+* Add `iOSPlatform` class.
+* Add `macOSPlatform` class.
+* Add `Darwin` class.
+* Add `iOS` class.
+* Add `macOS` class.
+* Add `DarwinVersion` class.
+* Move `OSPlatformImpl` to `DarwinPlatform`.
+* Remove `DarwinImpl` class.